removed node folder
