+++ cli/controllers/auth/register.py
