--- node/cli/index.coffee
@@ -1,7 +1,0 @@
-clis = [
-  require "./wallet"
-]
-
-module.exports.init = (program)->
-  for cli in clis 
-    cli.init(program)
