--- node/cli/wallet/create.coffee
@@ -1,2 +1,0 @@
-module.exports = (envs, options)->
-  console.log "create", envs
