--- node/cli/wallet/load.coffee
@@ -1,2 +1,0 @@
-module.exports = (envs, options)->
-  console.log "load", envs
