--- node/scripts/npm_prepare.sh
@@ -1,2 +1,0 @@
-npm prune;
-npm shrinkwrap;
