--- python/cli/server/__init__.py
@@ -1,7 +1,0 @@
-from flask import Flask
-
-
-class Server(Flask):
-  
-  def __init__(self):
-    super(Server, self).__init__(__name__)
